+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [#495](https://github.com/teambit/bit/issues/495) support adding files to imported components and ignoring existing files
 
+## [0.12.0-ext.11] - 2018-01-02
+
+- move `save-dependencies-as-components` flag from `bit import` command to be configurable in consumer bit.json
+- misc bugs for `bit remove`
+
 ## [0.12.0-ext.10] - 2017-12-27
 
 - add unsafe-perm for installing packages inside components